❗fix+ref: actions+preview: Continue large rework, also fixes some bugs
Fixes bug with previews without hotkey causing error due to them not
being resolved because they only PreviewChangeServerCalls in
ActionMenu were resolved.
Also fixes weird bug when chaining change-preview-window with
change-preview and having the command passed to --preview the same as
to change-preview causing change-preview not to be executed. The command
being different just by whitespace solves this issue, but instead I
chose to not use --preview at all.